@@ -2,18 +2,18 @@ source 'https://rubygems.org'
 ruby '2.1.5'
 
 # Application
-gem 'airbrake', '~> 4.0.0'
-gem 'rails', '~> 4.0.0'
+gem 'airbrake', '~> 4.0'
+gem 'rails', '~> 4.1.0'
 gem 'puma'
 
 # Views
-gem 'jquery-rails', '~> 3.1.2'
-gem 'tinymce-rails', '~> 4.1.6'
+gem 'jquery-rails', '~> 3.1'
+gem 'tinymce-rails', '~> 4.0'
 
 # Models
-gem 'pg', '~> 0.18.1'
-gem 'devise', '~> 3.0.0'
-gem 'acts-as-taggable-on', '~> 3.4.3'
+gem 'pg', '~> 0.18.0'
+gem 'devise', '~> 3.5.0'
+gem 'acts-as-taggable-on', '~> 3.0'
 
 group :development, :test do
   gem 'debase'
@@ -21,8 +21,8 @@ group :development, :test do
 end
 
 group :test do
-  gem 'rspec-rails', '~> 2.14.2'
-  gem 'factory_girl', '~> 4.5.0'
+  gem 'rspec-rails', '~> 2.14.0'
+  gem 'factory_girl', '~> 4.0'
   gem 'db-query-matchers', '~> 0.4.0'
 end
 
