@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150126154442) do
+ActiveRecord::Schema.define(version: 20160408222110) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
