@@ -73,22 +73,15 @@ describe PortfolioItem do
   end
 
   context "tagged lookup" do
-    before(:all) do
+    before do
       @model = PortfolioItem
       @model_sym = @model.name.underscore.to_sym
     end
 
     context ".active_taggings" do
       context "when empty" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -102,20 +95,13 @@ describe PortfolioItem do
       end
 
       context "when one tag contains only active items" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
           2.times do
             item = FactoryGirl.create(@model_sym)
             item.category_list.add("tag_one")
             item.save
           end
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -132,20 +118,13 @@ describe PortfolioItem do
       end
 
       context "when one tag contains only inactive items" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
           2.times do
             item = FactoryGirl.create(@model_sym, enabled: false)
             item.category_list.add("tag_one")
             item.save
           end
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -159,7 +138,7 @@ describe PortfolioItem do
       end
 
       context "when one tag contains some inactive items" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
 
           one = FactoryGirl.create(@model_sym)
@@ -168,13 +147,6 @@ describe PortfolioItem do
           two = FactoryGirl.create(@model_sym, enabled: false)
           two.category_list.add("tag_one")
           two.save
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -190,7 +162,7 @@ describe PortfolioItem do
       end
 
       context "when multiple tags contain the same items" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
 
           item = FactoryGirl.create(@model_sym)
@@ -198,13 +170,6 @@ describe PortfolioItem do
           item.category_list.add("tag_two")
           item.category_list.add("tag_three")
           item.save
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -222,7 +187,7 @@ describe PortfolioItem do
       end
 
       context "when things are really complicated" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
 
           @one   = FactoryGirl.create(@model_sym); @model.all
@@ -246,13 +211,6 @@ describe PortfolioItem do
           [@one, @two, @three, @four, @five, @six, @seven, @eight].each do |item|
             item.save
           end
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -295,7 +253,7 @@ describe PortfolioItem do
     end
 
     context ".tagged_items_from_taggings" do
-      before(:all) do
+      before do
         tagged_model_delete_all(@model)
       end
 
@@ -348,7 +306,7 @@ describe PortfolioItem do
     end
 
     context ".tags_from_taggings" do
-      before(:all) do
+      before do
         tagged_model_delete_all(@model)
       end
 
@@ -447,7 +405,7 @@ describe PortfolioItem do
     end
 
     context ".tag_names_from_taggings" do
-      before(:all) do
+      before do
         tagged_model_delete_all(@model)
       end
 
@@ -540,7 +498,7 @@ describe PortfolioItem do
     end
 
     context ".first_items_from_tags_from_taggings" do
-      before(:all) do
+      before do
         tagged_model_delete_all(@model)
       end
 
