@@ -21,7 +21,7 @@ describe ShopItem do
   end
 
   context "quantity" do
-    before(:all) do
+    before do
       @i = ShopItem.new
     end
 
@@ -42,7 +42,7 @@ describe ShopItem do
   end
 
   context "price" do
-    before(:all) do
+    before do
       @i = ShopItem.new
     end
 
@@ -120,22 +120,15 @@ describe ShopItem do
   end
 
   context "tagged lookup" do
-    before(:all) do
+    before do
       @model = ShopItem
       @model_sym = @model.name.underscore.to_sym
     end
 
     context ".active_taggings" do
       context "when empty" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once for Item, Tagging and Tag" do
@@ -149,20 +142,13 @@ describe ShopItem do
       end
 
       context "when one tag contains only active items" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
           2.times do
             item = FactoryGirl.create(@model_sym)
             item.category_list.add("tag_one")
             item.save
           end
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -179,20 +165,13 @@ describe ShopItem do
       end
 
       context "when one tag contains only inactive items" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
           2.times do
             item = FactoryGirl.create(@model_sym, enabled: false)
             item.category_list.add("tag_one")
             item.save
           end
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -206,7 +185,7 @@ describe ShopItem do
       end
 
       context "when one tag contains some inactive items" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
 
           one = FactoryGirl.create(@model_sym)
@@ -215,13 +194,6 @@ describe ShopItem do
           two = FactoryGirl.create(@model_sym, enabled: false)
           two.category_list.add("tag_one")
           two.save
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -237,7 +209,7 @@ describe ShopItem do
       end
 
       context "when multiple tags contain the same items" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
 
           item = FactoryGirl.create(@model_sym)
@@ -245,13 +217,6 @@ describe ShopItem do
           item.category_list.add("tag_two")
           item.category_list.add("tag_three")
           item.save
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -269,7 +234,7 @@ describe ShopItem do
       end
 
       context "when things are really complicated" do
-        before(:all) do
+        before do
           tagged_model_delete_all(@model)
 
           @one   = FactoryGirl.create(@model_sym, title: "Item One")
@@ -293,13 +258,6 @@ describe ShopItem do
           [@one, @two, @three, @four, @five, @six, @seven, @eight].each do |item|
             item.save
           end
-        end
-
-        before(:each) do
-          Rails.logger.debug "\nStarting test."
-        end
-        after(:each) do
-          Rails.logger.debug "Test finished.\n"
         end
 
         it "should hit the database only once per Item, Tagging and Tag" do
@@ -342,7 +300,7 @@ describe ShopItem do
     end
 
     context ".tagged_items_from_taggings" do
-      before(:all) do
+      before do
         tagged_model_delete_all(@model)
       end
 
@@ -395,7 +353,7 @@ describe ShopItem do
     end
 
     context ".tags_from_taggings" do
-      before(:all) do
+      before do
         tagged_model_delete_all(@model)
       end
 
@@ -494,7 +452,7 @@ describe ShopItem do
     end
 
     context ".tag_names_from_taggings" do
-      before(:all) do
+      before do
         tagged_model_delete_all(@model)
       end
 
@@ -587,7 +545,7 @@ describe ShopItem do
     end
 
     context ".first_items_from_tags_from_taggings" do
-      before(:all) do
+      before do
         tagged_model_delete_all(@model)
       end
 
