@@ -28,8 +28,8 @@ describe 'portfolio/_portfolio_categories' do
 
     it "should create a new list item for each category where enabled items exist" do
       render
-      rendered.should be_include("<li>" + link_to("cat 1", portfolio_item_category_path(URI::escape("cat 1")))  + "</li>")
-      rendered.should be_include("<li>" + link_to("cat 2", portfolio_item_category_path(URI::escape("cat 2")))  + "</li>")
+      rendered.should be_include("<li>" + link_to("cat 1", portfolio_item_category_path("cat 1"))  + "</li>")
+      rendered.should be_include("<li>" + link_to("cat 2", portfolio_item_category_path("cat 2"))  + "</li>")
     end
 
    it "should not create a new list item for each category where only disabled items exist" do
